新增功能:小程序查询浏览套餐,浏览菜品,分类 bug修复:jwt用户校验 代码优化:菜品分页查询逻辑复用

+++ sky-pojo/src/main/java/com/sky/dto/DishPageQueryDTO.java
@@ -13,8 +13,8 @@ public class DishPageQueryDTO implements Serializable {
 
     private String name;
 
-    //分类id
-    private Integer categoryId;
+    //分类id 修改为Long类型
+    private Long categoryId;
 
     //状态 0表示禁用 1表示启用
     private Integer status;
